requirements.txt
requests
PyGithub
pandas
